requirements.txt
pygame>=2.5.0
numpy>=1.24.0
faiss-cpu>=1.7.4
transformers>=4.40.0
torch>=2.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
